Fixed avalon reads
The avalon master cocotb driver removes the read address after 1 clock
cycle and sets it to x or u. This screws up the read process of the
implemented corsair avalon master module interface by giving out x's on
the read port.
It is fixed by clocking in the address into an internal read address and
use that for the read process.